[Feat] 카드리스트 레포 자료형 수정

+++ src/main/java/com/shinhan/soloplay/card/CardRepository.java
@@ -3,7 +3,7 @@ package com.shinhan.soloplay.card;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
-public interface CardRepository extends JpaRepository<CardEntity, Integer>, QuerydslPredicateExecutor<CardEntity>{
+public interface CardRepository extends JpaRepository<CardEntity, Long>, QuerydslPredicateExecutor<CardEntity>{
 	 
 
 }
